Refatoração e reestruturação da aplicação
