+++ src/main/java/com/br/alura/domain/curso/Curso.java
@@ -1,0 +1,21 @@
+package com.br.alura.domain.curso;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Table(name = "cursos")
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@EqualsAndHashCode(of = "id")
+public class Curso {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	private String nome;
+	private String categoria;
+
+}
