+++ src/main/java/com/br/alura/domain/topico/StatusTopico.java
@@ -1,4 +1,4 @@
-package com.br.alura.modelo;
+package com.br.alura.domain.topico;
 
 public enum StatusTopico {
