+++ src/main/java/com/br/alura/domain/topico/Topico.java
@@ -1,0 +1,45 @@
+package com.br.alura.domain.topico;
+
+import com.br.alura.domain.curso.Curso;
+import com.br.alura.domain.resposta.Resposta;
+import com.br.alura.domain.usuario.Usuario;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+
+@Entity
+@Table(name = "topicos")
+@Getter
+@Setter
+@AllArgsConstructor
+@EqualsAndHashCode(of = "id")
+@NoArgsConstructor
+public class Topico {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	private String titulo;
+	private String mensagem;
+
+	@Column(name = "data_criacao")
+	private LocalDateTime dataCriacao = LocalDateTime.now();
+	private StatusTopico status = StatusTopico.NAO_RESPONDIDO;
+	@ManyToOne
+	@JoinColumn(name = "autor_id")
+	private Usuario autor;
+	@ManyToOne
+	@JoinColumn(name = "curso_id")
+	private Curso curso;
+
+	@OneToMany(mappedBy = "topico")
+	private List<Resposta> respostas = new ArrayList<>();
+
+
+}
+
+
