+++ src/main/resources/application.properties
@@ -1,1 +1,10 @@
+###############CONFIG BD#####################
+spring.datasource.url=jdbc:mysql://localhost/forum_alura
+spring.datasource.username=root
+spring.datasource.password=
 
+###############CONFIG FORMAT CONSOLE##########
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+###############################################
